Add multi-reader, single-writer locks to kernel.
Any number of readers may be in the critical section at a time and writers
are mutually exclusive to all other threads. They are a better choice when
data is rarely modified but often read and multiple threads can safely
access it for reading.

For the moment, priority inheritance is not implemented. I do plan to get
around to doing so. It should be less of an issue when used appropriately
because of the multi-reader property.

This patch doesn't add anything to firmware/SOURCES. There is no reason
to do so until other code is pushed that uses them.

This code is tested and appears to work properly.

Change-Id: Idbe43dcd9170358e06d48d00f1c69728ff45b0e3