Update patch set 2
Patch Set 2:

+#if (CONFIG_PLATFORM & (PLATFORM_NATIVE|PLATFORM_HOSTED))

Well, this seems to do the trick. However, I've recently re-installed my laptop and switched it from 32 to 64-bit and the sim build spits out a ton of warnings like this:

warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

I think the best thing to do is avoid this whole casting debacle and do as the original Lua 5.1 embedding did, switch all the internal lua functions to use the "rockbox-native" int for any handling of files. It means more modifications to the lua core, but it avoids pointless/confusing casts.

Patch-set: 2