imx233/stfm1000: implement DRI interface, basic STFM1000 code and playback.
The tuner works (playback), one can tune and seek (a bit slow at the moment).
Recording does not work because I did not implement it yet. The code is a bit
of hack when it comes to playback because it uses the pcm interface directly
from the tuner code, and thus might break when switch in and out of radio
and to playback. A proper fix for this is to use the pcm mixer interface with
a custom channel, this TBD. RDS is completely unimplement at this point, it
requires to do all the RDS signal processing in software. Also the tuner audio
code forces mono mode currently. Having proper stereo from L+R and L-R needs
some filtering and RSSI analysis, otherwise it will be a disaster. The tuning
code still needs some improvement, some code is commented and hopefully will
be fixed later.