Add "NVIDIA" magic in our bootloader for the Sansa View.
Recovery mode of the original bootloader checks this magic. Now it is possible
to install our bootloader through the recovery mode. This was previously
possible only through the OF.

Benefits:
a) it's easier now to develop and test the images (mi4), because one don't have
   to go back to OF, in order to install another image to test,
b) it gives us a natural way of installing our bootloader in the future - the
   same method as in case of other pp players with flash memory (sa9200, sansa
   c200).

Change-Id: I530bff7d1fdb6c105b67a3eb0deb14253e1dbe74