RaaA: Android auto fit to any screen resolusion
This approach sets the Surface to a fixed size in respect of internal framebuffer resolution - include padding to keep image aspect ratio.

When render in a fixed-size buffer, the device hardware does the work of scaling the scene up (or down) to match the device's screen resolution.

I convert the input(x,y) position to match the internal framebuffer because scaling is only apply to Surface, not  MotionEvent.

theres a crash in text scrolling until I masked the rect update in lcd-scroll out.
Change-Id: I6c5a47a8374dd034c9a344d77beda8fe07a9dabc