random folder advance: add weights to folders
If there is a file named "rfa_config.txt" in a directory it will
be checked for two lines:

weight: <number>
recursive: yes/no

weight is some value 0-100 which represents the chance this directory
will be chosen.
recursive means this value will be used for all this directorys sub-dirs

You need to regenerate the list for the weights to take effect.

Change-Id: Ib421a46951e2f167cab9c2749e1e58cf83985289