Update patch set 5
Patch Set 5: Code-Review+1 Tested+1

Looks really good :)

Just tested with a 480x800 build on my 1080x1920 screen and while it looks obviously scaled it is very nice (I have cabbie again!).

It crashes consistantly switching between the cabbie theme and the default one (rockbox_failsafe) but not sure if this is because of your patch or not (need to verify).

Please add yourself to the docs/CREDITS file if you are not in there already.

Patch-set: 5