Update patch set 1
Patch Set 1:

This is just an RFC ... it is a workable version, though, and has already been beta-tested by the singer of my band (he was even able to seek to a certain part we wanted to practice;-).

Please see comments at the top of metronome2.c for usage info and changes to the normal metronome, to which this version may get merged in future.

It still works when started from plugin menu, but only offers fixed 4/4 beat, then. The interesting use case is opening a .tempo file with it. Example contents:

# pyramid.tempo: Increase speed
full: 2 1/1 120
half: 2 2/2 120
2tri: 2 3/3 120
quart: 2 4/4 120
4tri: 2 6/6 120
eights: 2 8/8 120
8tri: 2 12/12 120
16th: 2 16/16 120
16tri: 2 24/24 120
32th: 2 32/32 120
32tri: 2 48/48 120
64th: 2 64/64 120

A more conventional example:

# A song.
intro: 16 4/4 150
verse_1: 12 6/8 105
chorus_1: 8 4/4 150
verse_2: 12 6/8 105
chorus_2: 16 4/4 150
outro: 6 6/8 105

This is what I currently use it for: A metronome _track_ player. Actually, this might be a codec instead, integrated in the normal player. But: I need to be able to control time in terms of parts and bars (beat-based) instead of seconds (time-based). I guess the codec infrastructure doesn't allow that ... also the optical indication of beats bay be difficult.

Well ... comments welcome. This is the proof of concept.

Patch-set: 1