skin engine: Switch from offsets to actual pointers in all skin structs.
The skin engine was previously using offsets into the skins buflib
handles buffer. This was slightly error prone and bloated. Switch
back to using plain pointers and make sure to move them on the very
rare time when the handle moves.

Benefits of this include crahsing if there is a mistake instead of
randomly not working as correctly!

version 2:
 * Builds (hopefully) on everything except charcell
 * Finished removing the offset ptr macros!

Change-Id: I1d9555c5fe240b6d19d172edc6efc6e8b496b4f7