Added Music Quiz plugin.
Added music quiz plugin (similar to that found on ipods)
which generates a random playlist of tracks. These tracks are played
back with a list of possible track names displayed.
The user must correctly guess which track is playing to score points.
The quiz doesn't end until the user exists, new questions are generated
as the quiz continues.

The length of time each track plays for, the number of possible answers
displayed and whether those options hide as time progresses
are configurable.

Change-Id: Id40c98934e33ff8f9dc49846af724ebb1cb2ee88