Update patch set 2
Patch Set 2:

I don't want to be pain in the ass, but for discharging, with your values you never reach 100%, For example inverting highest value -> (4150+1097)*1024/5000 = 1075 which is out of range (1023+20). OF reaches 100% through the simple linear formula Bat_in_%=100*(ADC-860)/(1023-860).

Maybe it would be good to adapt the formula to the ADC range. This one is recalculated and gives exactly the same values for discharging (ADC range is 860..1023, so without surplus 0x14): mV = ADC*5932/1024-1780. Or we could change the values and adapt it to the existing formula; that would be: 3200,3493,3543,3602,3627,3653,3703,3771,3796,3868,3995.
mV values are needed only for runtime estimation (if it's enabled at all). What do you think about it all?

About PLL ifdefs: My bad, version from the first patch will be better. Someone on the very beginning assumed that ADC is only for the battery, which is not true. I want to add support for remote on YH92Xs, so i would like to have the code unified for all Samsungs and i dont want to add anything to readout.

Patch-set: 2
Label: Code-Review=0