Update patch set 3
Patch Set 3:

I've choosen the OF solution  - without addition depending on PLL (following the OF, Rockbox should substract 0x1a yet in case of SLEEP and DEFAULT freq, but i think it's not necessary for these two). This will help me add support for remote of YH92x or at least the source code will be more unified and cleaner.

Patch-set: 3