Update patch set 1
Patch Set 1: Code-Review-1 Tested-1

It compiles fine for iPod Classic, but does not work as it should. Pressing Menu (up) cancels the typing. Instead, you must scroll up to select the "up" option.

However, left, center, and right work as expected.

I see no way to go back in the event of an errant keypress.

It is a great idea, but needs some work.

Patch-set: 1