Speed up database.ignore/unignore processing with rewinddir()
The database code checks for "database.ignore" and "database.unignore"
using file_exists() twice in a row, requiring mutltiple internal scans
of the directory tree from the root. This is obviously inefficient.

g556 - "Rewrite filesystem code", implements rewinddir() natively but
doesn't enable it. This patch enables it natively and implements it for
the remaining hosted targets.

rewinddir() facilitates scanning for the special files in a first pass
then rewinding the directory to the beginning without reopening it
before the actual database scan of that directory begins. Also, only one
of the names has to be compared on each loop depending upon state.

Change-Id: I74bc4eabecceee09a6e62cae4070dc01cedb1a26