Update patch set 1
Patch Set 1: Tested+1

It works just as it should, but I have some suggestions:

 - Remove selections from the list as the song plays on
 - Add some kind of "lives" system to limit the amount of questions that can be missed OR a limit to the total amount of questions (both configurable)
 - Highlight the correct answer when the player does not choose one
 - Highlight the new high score when one is earned

Also, when using some themes (Redux, for example), the menu is shown in the wrong location. However, this is not a bug in the plugin but in the theming system, so you can do little about it.

It's looking really good!

Patch-set: 1