Update patch set 9
Patch Set 9:

The find_first_set_bit problem is a referencing issue because librockbox didn't have any use of it and that would cause it to get discarded. I just dumped usage of that for armv5 and above since __builtin_ctz() does the job and is pretty small. I fixed the typos in the AMS v2 stuff. My fuze v2 works nicely.

Patch-set: 9