Update patch set 13
Patch Set 13:

I messed up renaming in dircache when removing nulls from strings. Fixed that little oops. Threw in a better way to implement mempcpy quite cheaply by implementing as an alternate entry to memcpy in the ASM files. I think I'll be done with it for now unless something is uncovered. I would like to push support code ahead of the remainder of this whether or not it's currently used in order to cut back on the patch size.

Patch-set: 13
Label: Code-Review=0