Update patch set 14
Patch Set 14:

This one covers something I realized when in IRC. If the file has an uncommitted sector on the end that has never been part of the cluster chain, committing it could fail if the device is full (of course). The directory entry and the cluster chain would then disagree as to the file size because the size in the descriptor is already larger than what's been written but nothing more can be written. The change here will perform a truncation in that case to get the two back into agreement (you were gonna lose it anyway).

Patch-set: 14