Add mempcpy implementation
A GNU extension that returns dst + size instead of dst. It's a nice
shortcut when copying strings with a known size or back-to-back blocks
and you have to do it often.

May of course be called directly or alternately through
__builtin_mempcpy in some compiler versions.

For ASM on ARM and m68k, it is implemented as an alternate entrypoint
to memcpy which adds minimal code and overhead. This may be doable for
the others.

Change-Id: I4cbb3483f6df3c1007247fe0a95fd7078737462b