Update patch set 2
Patch Set 2: -Code-Review

Just to clarify: I said that because firstly, there was software volume control in the dsp code but it removed and implemented directly in the PCM layer. It suffers from two problems: 1) excess latency 2) doesn't operate on all PCM output because not everything uses DSP processing. Basically, you patch it into the hardware driver or what substitutes for one. There are several examples in the codebase besides what I mentioned. I'd use the unbuffered version on a hosted target (see pcm-sdl.c for the latter and buffered on a native one (see hifiman and jz47xx for the former).

Patch-set: 2