Update patch set 3
Patch Set 3:

the second dsp, fatigue reduction, is actually not for voice but for music playback. I did some study of research papers and evaluated with frequency of sounds(human hearing equal loudness), found out that the sounds between 2khz~7khz can trigger TTS: change the threshold of hearing. so listener would start to feel the bass and treble are weaken.
and add some gain in lower band - which will generate overtone, and masked the higher frequency.
even worst, some might not know the muddy bass is because the overtone and try to add gain to treble, for clarity, somewhere between 2khz to 10khz. where TTS could happen again if the playback volume is high.

Patch-set: 3