Update patch set 3
Patch Set 3:

The EQ precut is PGA gain, the one for BBE(pbe) is the overall gain of the filter. they apply in different stages, cannot be the same. try different combination of eq precut and BBE precut yourself. they have different sound signature,and not interchangeable.

Patch-set: 3
Label: Code-Review=0