Update patch set 4
Patch Set 4:

Also, I wouldn't be caching the those values in global variables like that. Configure can be called for other reasons. I make no promise not to break stuff in the future if it doesn't follow the intended interface. Use existing implementations as your examples of what is correct.

Patch-set: 4