Update patch set 1
Patch Set 1:

Since the deadzone patch builds and runs under the sim, and I've been running it for months on my fuze+ with nothing but gratitude, I think it's a good candidate to be merged.  Here's a link to a PDF of the compiled manual.  The entry is in "General Settings > System > Touchpad".

http://creativepc.iolab.com/rockbox-sansafuzeplus.pdf

Patch-set: 1