Update patch set 1
Patch Set 1:

definition SMARTPHONE_UI_STYLE is not needed by the port. It is something that came up with Ilias builds when only allowed power button to enable backlight. Everyone who was using rockbox on DX50 it got used to it (including me) and I wanted to include it in a proper way. I first wanted to take it out for the review but in the end it was overdue to do the commit and this commit was not really meant as a final commit to be accepted. All targets (including DX50&90) should compile as normal when SMARTPHONE_UI_STYLE is not defined.

Patch-set: 1