Update patch set 1
Patch Set 1:

Patching the OF is ok, we do that on many targets already.  Ideally though we should have a tool that does it automatically for the user (for example on the Sansa Clip we ask the user for a firmware file, and the tool adds rockbox to it).  For now this is not important however, so I would not worry too much about it just yet.

I don't think you should include UI changes in this commit.  I would split them as a second gerrit task as they will probably need to be reviewed by a different set of people.  In the meantime, since you are doing custom builds anyway, it will be very easy for you to add it back in your local tree.  My main concern is making sure that the DX port doesn't drift far enough from our source code that it can't be integrated in the future.  Once its in, we can discuss UI changes, and if other targets should have access to that feature as well.

Patch-set: 1
Label: Code-Review=0