Update patch set 9
Patch Set 9:

What is the canonical form of the PBE as y[n] = ...? Really the temp buffer is huge and shouldn't need to be large at all, maybe even stackable. The sample count that it must technically deal with is 1 to INT_MAX. It also needs to do mono (which just needs an outer channel loop). I feel like this could be done more expediently as well (just a hunch).

Patch-set: 9
Label: Code-Review=0