Update patch set 2
Patch Set 2:

I'm new to git and gerrit so I'm surprised that everithing came out I wanted to :-): 
The second patch set is free of UI changes and is the plain DX50 & DX90 port. I've also did some whitespace cleanup and found a bug on the DX90 port that I removed.
Unfortunately headwhacker did not give me any feedback of the build I've sent him yet for the DX90.

Patch-set: 2