Update patch set 16
Patch Set 16:

Thanks for the detail report!
I was already expecting the semi responsive situation would occur on native targets after remove dsp_filter(in afr and pbe) from IRAM. but it is amazing that take so long before it actually happen. Now we require a developer who has native target work on  the test to check which variable and dsp_filter needs to be placed in IRAM. (Since I can only test on host targets)

Patch-set: 16
Label: Code-Review=0