Update patch set 1
Patch Set 1: Code-Review+1 Tested+1

Functions as expected here.

Tested on iPod Classic 6/7G (including the CE-ATA variant), with both platter based and solid state disks, and on iPod Nano 2G.

There is a fairly reasonable increase in throughput performance (10~20% from my tests) on the iPod Classic, but the improvement in throughput I saw on the iPod Nano 2G could be attributed to standard deviation (1~3%).

I am unsure as to whether or not this is expected, or should be considered an additional bonus.

I would be able to re-test and supply test results, if such information would indeed be desirable.

Well done Marcin, and Michael.
Much appreciated.

Patch-set: 1
Label: Code-Review=+1
Label: Tested=+1