usb-designware: New USB driver for Synopsys DesignWare USB OTG core.
The new driver is enabled on iPod Nano 2G and iPod Classic/6G,
so USB storage finally works again on those targets.
This is rewrite of the driver to get rid of bitfields. It tries
to be as close as possible to original TheSeven's work.
Original code is left in comments to easy validation of rewrite.
This is WIP not intended to be commited without major cleanup!
Change-Id: Iab63c3f7d8f4a75351af1a5282f9d1b231ae7bd1