Update patch set 2
Patch Set 2: Tested+1

I'm only able to test this on the iPod 6G, so I'm not even sure it doesn't break the build for other platforms; apologies if it does, and advice for how to fix the patch would be appreciated if so.

There was some concern on IRC when I originally proposed this that it might cause the battery to not charge (or even to drain) if such an option were added, but it doesn't seem to cause any issue for me.  It would be a good idea to add a warning if the patch is applied to other platforms that this might occur.

Patch-set: 2