Update patch set 2
Patch Set 2: -Tested

a) Using the device in disk mode is very slow if it spins down constantly.  The option to disable spindown entirely is not a substitute, because then you have to remember to turn spindown off before plugging the device in, and remember to turn it back on after unplugging it; users may forget.

b) I knew this would be an issue, but I didn't have a good way to test the patch on other devices.  If I fix this, would you (or someone else) be able to help?

c) No problem, I can write one.

Patch-set: 2
Reviewer: Gerrit User 1000423 <1000423@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Tested=0