Update patch set 9
Patch Set 9:

this one now works (problem was a syntax error of a warning cleanup). There have been code changes in the meantime that make the start of playback very slow when database scan is active. It also seems that the hack with DEFAULT_STACK_SIZE of patchset 5 is not necessary anymore (but I left it).

Patch-set: 9