Update patch set 9
Patch Set 9:

I tried out this version today and it works well.
In the meantime I got a feedback from a DX90 tester. He said that the app does not start. Status was patchset4. It depends on Michael if he wants to do the commit anyway or not. In my opinion it can be done and the fix for DX90 added later. In the meantime i will try to find the bug. My tester (deho@head-fi.org) has now everything set up to do tests and is willing to help.

Patch-set: 9
Label: Code-Review=0