Update patch set 1
Patch Set 1:

Does this have any noticable effect on ATA performance?
(Boot time, dircache/database build time, etc.)

If not, I'm perfectly fine with it.
In fact this is how emCORE does it, just that emCORE generally has more fine-grained yielding and preemptive multitasking.

I think the reason for not yielding in this loop in Rockbox were concerns about over-yielding in some threads, and other drivers not yielding in that place either (at that time).

There's no target-specific reason to not do this, so I don't have an opinion on this as the target maintainer. However someone with more in-depth knowledge about Rockbox's threading/scheduling should have a look at this.

Patch-set: 1
Reviewer: Gerrit User 1000351 <1000351@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0