Update patch set 1
Patch Set 1:

(1 comment)

The idea behind the keymap was to use the REC switch (YH920) as "switch on-switch off" combination, so it would be the same as a long REC press on YH820. But I must admit that from a user's point of view it's more like an unexpected switch-misuse.

Unfortunately a switched REC state can almost never be used, because it will break most keymaps (like mentioned in the inline comment).

After some meditation I come to the conclusion that the best thing would be to ignore the REC switch state in the keymaps and omit BUTTON_REC (on YH9xx), but to introduce two (virtual) buttons that fire a single button event only when the REC switch changes (e.g. BUTTON_REC_STATE_CHANGES_TO_ON resp. OFF)). That way all benefits of a REC switch could be used without breaking any keymaps. The target could even be started and used if REC is switched on accidentally.

Patch-set: 1
Reviewer: Gerrit User 1000293 <1000293@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0