move binaries for iBasso DX50 & DX90 to internal drive (/sdcard)
This patchset changes three things:
* binaries, libs, all rockbox files AND user settings are now in the
  same folder on the fat32 partition in the folder "/mnt/sdcard/.rockbox"
* to have rockbox_dir in /.rockbox, the dualboot selector app now
  creates a symbolic link at every start to /mnt/sdcard/.rockbox
* the boot selector app shows an error picture when no rockbox binary is
  in the internal drive. When rockbox crashes (this happens when usb
  mass storage is active) a USB-connected image is shown.

The first part of this patch will simplify the update procedure a lot.
No more patching of the OF will be required for an update, only once
for the initial installation (pending). Now the folder structure is
equal to the the one of any regular target.

Change-Id: I37ac39f1635b4e65a0695d5bd2418325e9ef0279