Update patch set 3
Patch Set 3:

a) this patch removes the HAVE_SPECIAL_DIRS define, so rockbox is in the rbdir and we avoid many problems if rbdir is "/.rockbox". This symbolic link is only in the ram filesystem.
b) as it is it does not work. I will check the function you've mentioned, where the call for it is with the option for android.

Patch-set: 3