application: Map /.rockbox to the actual ROCKBOX_DIR on config load
application ports may not use /.rockbox for various reasons (one being that / is a
read-only filesystem). To make themes loadable anyway introduce a special case
for theme (and other) files that are attempted to be loaded from /.rockbox.

This is one way, it will not map from arbiratry dirs to /.rockbox.

Change-Id: I8e7515841d01a0ef28d89bc17da14f405ffb2b60