Update patch set 4
Patch Set 4:

Patch Set 4: in this stage I would prefer to do the commit.
Codecs are now named as X.codecs not as libX.so, therefore the buildzip.pl does not need to be changed (thanks a lot to Thomas). +some code cleanup in ibassodualboo.c thanks to Brendanus (thanks to him also for testing Patchset 3)

Patch-set: 4