Update patch set 2
Patch Set 2:

(19 comments)

First, great work. There are lots of whitespace errors (almost everywhere). Now the licensing - from the technical point of view the plugin is derived work of rockbox as it is statically linked with rb stuff. This implies that the code needs to be GPLv2+ compatible. I understand the situation of the code and hence if you want it to be merged the licensing terms MUST be GPLv2+ compatible and not 'in the spirit of GPL'. There should be explicit statement that resource files are external and covered by other license. We also have real name policy which means that original author should be known and properly credited. We also insist on providing manual entry.

Patch-set: 2