"Modern User Interface": Turn off Backlight and Touchscreen with Power-Button (in WPS)
Depends on: http://gerrit.rockbox.org/r/#/c/1043/4

This Patch changes the behaviour of buttons and backlight if set up correctly (here
applied for the targets DX50, DX90 and D2).
When in WPS, backlight can only be switched on with a button defined in button-target.h.
The keymap has ben modified that the same button causes the backlight to turn off.

The idea behind this patch is that on some touchscreen targets it is a need  to turn
off backlight and with it touchscreen to use its hardware playback control buttons
(like "skip" or volume change) blind in a pocket without doing unwanted actions by
accidentally touching the touchscreen when player has been put into the pocket before
the backlight timeout has expired.

Patch has no effect on targets that have not the define MODERN_UI.
When defined, setting "ignore first keypress" is completely removed from menu
because it doesn't make sense anymore with MODERN_UI.

Change-Id: I5457bdb1245e2985e0dadfb78b7f47bac8292785