"Modern User Interface": Toggle Backlight with a button in WPS for on some Touchscreen Targets
This Patch changes the behaviour of buttons and backlight if set up correctly (here
applied for the targets DX50, DX90 and D2).
When in WPS, backlight can only be switched on with a button defined in button-target.h.
The keymap has ben modified that the same button causes the backlight to turn off.
The second part of the patch is that "cancel" in the main menu does the action of resuming
the playback (brings you to WPS).
The idea behind this patch is that it should be easy to turn off backlight and with it
touchscreen to use touchscreen-players's hardware buttons (like "skip" or volume change)
blind in a pocket without doing unwanted actions by accidentally touching the touchscreen.

Patch has no effect on targets that have not the define MODERN_UI. With patch applied,
setting "ignore first keypress" has no effect anymore.

Change-Id: I5457bdb1245e2985e0dadfb78b7f47bac8292785