"Modern User Interface": Toggle Backlight with a button in WPS
This Patch changes the behaviour of buttons and backlight if set up correctly (here
applied for the targets DX50 and DX90).
When in WPS, backlight can be switched off an on with a button defined in button-target.h.
The second part of the patch is that "cancel" in the main menu does the action of resuming
the playback (brings you to WPS).
The idea behind this patch is a) that WPS should be the "main screen" not the menu
and b) that backlight should easily be switchable because we do that every day x times
on our smartphone and got so used to this.

Patch has no effect on targets that have not the define MODERN_UI. With patch applied,
setting "ignore first keypress" has no effect anymore.

Change-Id: I5457bdb1245e2985e0dadfb78b7f47bac8292785