Update patch set 2
Patch Set 2:

Please, don't PLA-fy the code!
IMO a good plugin deserves optimized keymaps
(and of course I can help out for some targets :) ).
PLA is a convenience for quick-and-simple plugins,
but it's not as straightforward as one may think,
and surely will degrade usability on some targets.

Apart from this you'll give away the chance to easily 
improve keymaps for individual targets and to 
add actions on targets with sufficient buttons
(e.g. game sound volume)

Patch-set: 2
Label: Code-Review=0