Update patch set 1
Patch Set 1: Code-Review-1

I don't like the change honestly. The 'rockbox way' is to turn off backlight when hold is engaged. You can define SW implementation of this (combo or long press or whatever) which will have the same effect with added bonus of locked keys (That is how this works on phones anyway). fuze+ or clip can serve as example.
Describing WPS as 'main screen' is questionable but we do have the setting to start last played track on boot so this jumps straight to WPS, we have action to move from WPS to root menu here and forth. I don't see the point of reinventing the weel.

Patch-set: 1
Reviewer: Gerrit User 1000020 <1000020@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=-1