Update patch set 4
Patch Set 4:

At the moment not because the patch does not include the part that Thomas made his patch for. I tested his patch with an arbitary ROCKBOX_DIR and themes where not loading as they should (what Thoma's patch is actually for).
I would prefer the patch as it is right now but of course I will add the changing Thomas wants when I'm sure it works.
The reason why I prefer the way it is, is because DX90 and DX50 port will be the only port not running from "/.rockbox" (exept the HAVE_SPECIAL_DIRS ports). So I fear that it will be the first port to not work anymore every time a code changes come that affect the file structure. I did not understand why Thomas does not like my symbolic link, but he's the senior developer, I will do what he wants.

Patch-set: 4